requirements.txt
jupyter-book
matplotlib
numpy
pyppeteer
ipython
plotly
pandas

# data processing reqs
python-dateutil
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib